// src/app/shared/models/user.model.ts
import { IOrder } from "./order.model";

export interface IUser {
  access_url?: string;
  address_1?: string;
  address_2?: string;
  address_3?: string;
  age?: string;
  city?: string;
  disable_foreign_donation?: number;
  display_type?: string;
  email?: string;
  entity_type?: string;
  extension?: string;
  fname?: string;
  full_name?: string;
  gender?: string;
  id?: number;
  is_subscribed?: number;
  last_communication_date?: string;
  lname?: string;
  no_80g?: string;
  phone_1?: string;
  phone_2?: string;
  pincode?: number;
  priority?: string;
  status_flag?: string;
  updation_time?: string;
  user_details_id?: number;
  user_role?: string;
  username?: string;
  website_1?: string;
  website_2?: string;
  pan_number?: any;
  avtar?: {
    cdn_path: string,
    entity_type?: string,
    entity_type_id?: number
    file_name?: string,
    media_type?: string,
    path?: string,
  };
  donor: IOrder[];
  subscribed?: {
    amount?: number;
    attempt?: number;
    campaign_id?: number;
    campaign_cause?: string;
    creation_time?: string;
    currency?: string;
    donor_entity_details_id?: number;
    end_date?: string;
    id?: number;
    next_payment_date?: string;
    ngo_entity_details_id?: number;
    plan_type?: string;
    split_amount?: string;
    start_date?: string;
    status_flag?: number;
    tenure?: number;
    update_time?: string;
    payment_gateway?: string;
    payment_mode?: string;
  };
  aggdonationall: {
    donor_entity_details_id: number,
    funds: number,
    campaigns: number,
    donations: number,
    avg_donation: number
  };
  sip_subscribed?: {
    action_date?: string,
    amount?: number,
    attempt?: number,
    campaign_id?: number,
    creation_time?: string,
    currency?: string,
    donor_entity_details_id?: number,
    end_date?: string,
    id?: number,
    next_payment_date?: string,
    ngo_entity_details_id?: number,
    payment_gateway?: string,
    payment_mode?: string,
    payment_status?: number,
    plan_type?: string,
    split_amount?: string,
    start_date?: string,
    status_flag?: number,
    tenure?: number,
  };
  lastorder: {
    campaign_id: number;
    cause_id: number;
    checksum: string;
    creation_date: string;
    creator_entity_details_id: number;
    currency: number;
    customer_id: string;
    device: string;
    donated_amount: number;
    donated_amount_local: number;
    donated_amount_return: number;
    donated_amount_usd: number;
    entity_details_id: number;
    exp_month: number;
    exp_year: number;
    funding: string;
    gateway: string;
    id: number;
    is_anonymous: number;
    is_indian: number;
    iso_currency: string;
    last4: string;
    merchant_key: string;
    network: string;
    order_id: string;
    parent_entity_details_id: number;
    parent_order_id: number;
    payment_gateway: string;
    payment_mode: string;
    payment_mode_value: string;
    private_key: string;
    rewards_id: number;
    subscription_id: number;
    tip_amount: number;
    tip_amount_inr: number;
    tip_amount_local: number;
    total_amount_charged: number;
    updation_date: string;
    upi_id: string;
  };
  active_subscription?: any;
  listsubscriptions?: {
    animals?: any,
    education?: any,
    medical?: any,
    memorial?: any
  };
}